pyproject.toml
[project]
name = "rwb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.2.15",
    "biopython>=1.85",
    "duckduckgo-search>=8.0.0",
    "fastrtc[stt,tts,vad]>=0.0.20",
    "ipython>=9.1.0",
    "kokoro>=0.9.4",
    "librosa>=0.11.0",
    "lxml-html-clean>=0.4.2",
    "markdown>=3.8",
    "newspaper4k>=0.9.3.1",
    "numpy>=2.2.4",
    "ollama>=0.4.7",
    "pip>=25.0.1",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "pygame>=2.6.1",
    "pyside6>=6.9.0",
]
